Move list extensions into companion object
