--- arrow-libs/core/arrow-core-data/src/main/kotlin/arrow/core/List.kt
@@ -1,16 +1,0 @@
-package arrow.core
-
-import arrow.typeclasses.Monoid
-import arrow.typeclasses.Semigroup
-import kotlin.collections.plus as _plus
-
-fun <A> Semigroup.Companion.list(): Semigroup<List<A>> =
-  Monoid.list()
-
-fun <A> Monoid.Companion.list(): Monoid<List<A>> =
-  ListMonoid as Monoid<List<A>>
-
-object ListMonoid : Monoid<List<Any?>> {
-  override fun empty(): List<Any?> = emptyList()
-  override fun List<Any?>.combine(b: List<Any?>): List<Any?> = this._plus(b)
-}
